remove useless comment and add `ID`/`PW` field

+++ blueprint/blog.py
@@ -6,32 +6,28 @@ from werkzeug.exceptions import abort
 from .auth import login_required
 from script.db import get_db
 
-''' Step 0. BluePrint(청사진; 도면) 설정
- => 이 도면을 사용해서 만들면, (도메인)/blog/(설정한 url)으로 HTTP 요청을 받을 수 있음
- ex> @bp.route('/post') => https://www.songnae.com/blog/post
-'''
 bp = Blueprint('blog', __name__)
 
-''' Step 1. HTTP 요청을 처리할 URL과 방식을 설정
- => 방식은 아래 두번째 함수에 있듯, `methods=('GET', 'POST')` 이런 식으로 설정할 수 있음
- 방식 생략시 기본값 = GET 요청
-'''
 @bp.route('/')
 def index():
     db = get_db()
     posts = db.execute(
-        'SELECT p.id, title, body, created, author_id, username'
-        ' FROM post p JOIN user u ON p.author_id = u.id'
+        'SELECT id, title, content, author, PW, created'
+        ' FROM post'
         ' ORDER BY created DESC'
     ).fetchall()
     return render_template('blog/index.html', posts=posts)
 
 @bp.route('/create', methods=('GET', 'POST'))
-@login_required
 def create():
     if request.method == 'POST':
         title = request.form['title']
-        body = request.form['body']
+        content = request.form['content']
+        ID = request.form['ID']
+        print(f'{ID = }')
+        PW = request.form['PW']
+        print(f'{PW = }')
+
         error = None
 
         if not title:
@@ -42,52 +38,50 @@ def create():
         else:
             db = get_db()
             db.execute(
-                'INSERT INTO post (title, body, author_id)'
-                ' VALUES (?, ?, ?)',
-                (title, body, g.user['id'])
+                'INSERT INTO post (title, content, author, PW)'
+                ' VALUES (?, ?, ?, ?)',
+                (title, content, ID, PW)
             )
             db.commit()
             return redirect(url_for('blog.index'))
 
     return render_template('blog/create.html')
 
-def get_post(id, check_author=True):
+def get_post(id):
     post = get_db().execute(
-        'SELECT p.id, title, body, created, author_id, username'
-        ' FROM post p JOIN user u ON p.author_id = u.id'
-        ' WHERE p.id = ?',
+        'SELECT id, title, content, author, PW, created'
+        ' FROM post'
+        ' WHERE id = ?',
         (id,)
     ).fetchone()
 
     if post is None:
         abort(404, f"Post id {id} doesn't exist.")
 
-    if check_author and post['author_id'] != g.user['id']:
-        abort(403)
-
     return post
 
 @bp.route('/<int:id>/update', methods=('GET', 'POST'))
-@login_required
 def update(id):
     post = get_post(id)
 
     if request.method == 'POST':
-        title = request.form['title']
-        body = request.form['body']
+        print(request.form['value'])
+        content = request.form['content']
+        PW = request.form['PW']
+
         error = None
 
-        if not title:
-            error = 'Title is required.'
-
+        if post['PW'] != PW:
+            error = 'Wrong Password'
+        
         if error is not None:
             flash(error)
         else:
             db = get_db()
             db.execute(
-                'UPDATE post SET title = ?, body = ?'
+                'UPDATE post SET content = ?'
                 ' WHERE id = ?',
-                (title, body, id)
+                (content, id)
             )
             db.commit()
             return redirect(url_for('blog.index'))
@@ -95,10 +89,19 @@ def update(id):
     return render_template('blog/update.html', post=post)
 
 @bp.route('/<int:id>/delete', methods=('POST',))
-@login_required
 def delete(id):
-    get_post(id)
-    db = get_db()
-    db.execute('DELETE FROM post WHERE id = ?', (id,))
-    db.commit()
+    post = get_post(id)
+    PW = request.form['PW']
+
+    error = None
+
+    if post['PW'] != PW:
+        error = 'Wrong Password'
+
+    if error is not None:
+        flash(error)
+    else:
+        db = get_db()
+        db.execute('DELETE FROM post WHERE id = ?', (id,))
+        db.commit()
     return redirect(url_for('blog.index'))
